feat: implement the logic of getting manufacturing info about a product

+++ services/backend/product-service/src/main/java/com/veteroch4k/product/controllers/DrawingController.java
@@ -32,7 +32,7 @@ public class DrawingController {
   }
 
   @GetMapping("/{id}")
-  public ResponseEntity<Drawing> getDrawing(@PathVariable long id) {
+  public ResponseEntity<Drawing> getDrawing(@PathVariable int id) {
 
     return drawingRepository
         .findById(id)
